SONAR-6615 add in table 'manual_measures' column 'compunent_uuid'
