+++ server/sonar-server/src/main/java/org/sonar/server/measure/custom/persistence/CustomMeasureDao.java
@@ -25,8 +25,8 @@ import java.util.List;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 import org.sonar.api.server.ServerSide;
-import org.sonar.core.custommeasure.db.CustomMeasureDto;
-import org.sonar.core.custommeasure.db.CustomMeasureMapper;
+import org.sonar.core.measure.custom.db.CustomMeasureDto;
+import org.sonar.core.measure.custom.db.CustomMeasureMapper;
 import org.sonar.core.persistence.DaoComponent;
 import org.sonar.core.persistence.DaoUtils;
 import org.sonar.core.persistence.DbSession;
