+++ sonar-core/src/main/java/org/sonar/core/persistence/DatabaseVersion.java
@@ -34,7 +34,7 @@ import org.sonar.api.server.ServerSide;
 @ServerSide
 public class DatabaseVersion {
 
-  public static final int LAST_VERSION = 917;
+  public static final int LAST_VERSION = 919;
 
   /**
    * List of all the tables.n
