+++ sonar-core/src/main/java/org/sonar/core/persistence/MyBatis.java
@@ -53,8 +53,8 @@ import org.sonar.core.component.db.SnapshotMapper;
 import org.sonar.core.computation.db.AnalysisReportDto;
 import org.sonar.core.computation.db.AnalysisReportMapper;
 import org.sonar.core.config.Logback;
-import org.sonar.core.custommeasure.db.CustomMeasureDto;
-import org.sonar.core.custommeasure.db.CustomMeasureMapper;
+import org.sonar.core.measure.custom.db.CustomMeasureDto;
+import org.sonar.core.measure.custom.db.CustomMeasureMapper;
 import org.sonar.core.dashboard.ActiveDashboardDto;
 import org.sonar.core.dashboard.ActiveDashboardMapper;
 import org.sonar.core.dashboard.DashboardDto;
