+++ sonar-core/src/main/resources/org/sonar/core/measure/custom/db/CustomMeasureMapper.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.sonar.core.custommeasure.db.CustomMeasureMapper">
+<mapper namespace="org.sonar.core.measure.custom.db.CustomMeasureMapper">
   <sql id="selectColumns">
     m.id,
     m.metric_id as metricId,
     m.resource_id as componentId,
+    m.component_uuid as componentUuid,
     m.value,
     m.text_value as textValue,
     m.user_login as userLogin,
@@ -37,12 +38,12 @@
 
   <insert id="insert" parameterType="CustomMeasure" useGeneratedKeys="true" keyColumn="id" keyProperty="id">
     INSERT INTO manual_measures (
-    metric_id, resource_id, value, text_value, user_login, description, created_at, updated_at
+    metric_id, resource_id, component_uuid, value, text_value, user_login, description, created_at, updated_at
     )
     VALUES (
-    #{metricId, jdbcType=INTEGER}, #{componentId, jdbcType=INTEGER}, #{value, jdbcType=DOUBLE},
-    #{textValue, jdbcType=VARCHAR}, #{userLogin, jdbcType=VARCHAR},#{description, jdbcType=VARCHAR},
-    #{createdAt, jdbcType=BIGINT}, #{updatedAt, jdbcType=BIGINT}
+    #{metricId, jdbcType=INTEGER}, #{componentId, jdbcType=INTEGER}, #{componentUuid, jdbcType=VARCHAR},
+    #{value, jdbcType=DOUBLE}, #{textValue, jdbcType=VARCHAR}, #{userLogin, jdbcType=VARCHAR},
+    #{description, jdbcType=VARCHAR}, #{createdAt, jdbcType=BIGINT}, #{updatedAt, jdbcType=BIGINT}
     )
   </insert>
