+++ sonar-core/src/main/resources/org/sonar/core/persistence/rows-h2.sql
@@ -341,6 +341,8 @@ INSERT INTO SCHEMA_MIGRATIONS(VERSION) VALUES ('914');
 INSERT INTO SCHEMA_MIGRATIONS(VERSION) VALUES ('915');
 INSERT INTO SCHEMA_MIGRATIONS(VERSION) VALUES ('916');
 INSERT INTO SCHEMA_MIGRATIONS(VERSION) VALUES ('917');
+INSERT INTO SCHEMA_MIGRATIONS(VERSION) VALUES ('918');
+INSERT INTO SCHEMA_MIGRATIONS(VERSION) VALUES ('919');
 
 INSERT INTO USERS(ID, LOGIN, NAME, EMAIL, CRYPTED_PASSWORD, SALT, CREATED_AT, UPDATED_AT, REMEMBER_TOKEN, REMEMBER_TOKEN_EXPIRES_AT) VALUES (1, 'admin', 'Administrator', '', 'a373a0e667abb2604c1fd571eb4ad47fe8cc0878', '48bc4b0d93179b5103fd3885ea9119498e9d161b', '1418215735482', '1418215735482', null, null);
 ALTER TABLE USERS ALTER COLUMN ID RESTART WITH 2;
