+++ sonar-core/src/main/resources/org/sonar/core/persistence/schema-h2.ddl
@@ -282,6 +282,7 @@ CREATE TABLE "MANUAL_MEASURES" (
   "ID" BIGINT NOT NULL GENERATED BY DEFAULT AS IDENTITY (START WITH 1, INCREMENT BY 1),
   "METRIC_ID" INTEGER NOT NULL,
   "RESOURCE_ID" INTEGER,
+  "COMPONENT_UUID" VARCHAR(50),
   "VALUE" DOUBLE,
   "TEXT_VALUE" VARCHAR(4000),
   "USER_LOGIN" VARCHAR(255),
@@ -624,6 +625,8 @@ CREATE INDEX "WIDGET_PROPERTIES_WIDGETS" ON "WIDGET_PROPERTIES" ("WIDGET_ID");
 CREATE INDEX "PROPERTIES_KEY" ON "PROPERTIES" ("PROP_KEY");
 
 CREATE INDEX "MANUAL_MEASURES_RESOURCE_ID" ON "MANUAL_MEASURES" ("RESOURCE_ID");
+
+CREATE INDEX "MANUAL_MEASURES_COMPONENT_UUID" ON "MANUAL_MEASURES" ("COMPONENT_UUID");
 
 CREATE INDEX "PROJECTS_KEE" ON "PROJECTS" ("KEE", "ENABLED");
